Create README.md
This is my 6th lab work. This is the telegram bot of my study group's schedule. I used the following libraries: psycopg2 - to connect to PostgreSQL, perform SQL queries and other database operations; telebot - to create a python bot for telegram.  Module: datetime - to work with date and time in python.